+++ component/src/main/java/io/siddhi/extension/io/grpc/util/GrpcUtils.java
@@ -17,6 +17,7 @@
  */
 package io.siddhi.extension.io.grpc.util;
 
+import io.siddhi.core.exception.SiddhiAppRuntimeException;
 import io.siddhi.query.api.exception.SiddhiAppValidationException;
 
 import java.util.ArrayList;
@@ -24,6 +25,9 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
+/**
+ * Class to hold the static util methods needed
+ */
 public class GrpcUtils {
     public static String getServiceName(String path) {
         List<String> urlParts = new ArrayList<>(Arrays.asList(path.split(GrpcConstants.PORT_SERVICE_SEPARATOR)));
@@ -56,5 +60,26 @@ public class GrpcUtils {
         List<String> urlParts = new ArrayList<>(Arrays.asList(path.split(GrpcConstants.PORT_SERVICE_SEPARATOR)));
         urlParts.removeAll(Collections.singletonList(GrpcConstants.EMPTY_STRING));
         return urlParts.size() == 3;
+    }
+
+    public static String[] extractHeaders(String headerString, String[] requestedTransportPropertyNames) {
+        String[] headersArray = new String[requestedTransportPropertyNames.length];
+        String[] headerParts = headerString.split(GrpcConstants.STRING_COMMA);
+        for (String headerPart: headerParts) {
+            String cleanA = headerPart.replaceAll(GrpcConstants.STRING_INVERTED_COMMA, GrpcConstants.EMPTY_STRING);
+            cleanA = cleanA.replaceAll(GrpcConstants.STRING_SPACE, GrpcConstants.EMPTY_STRING);
+            String[] keyValue = cleanA.split(GrpcConstants.PORT_HOST_SEPARATOR);
+            for (int i = 0; i < requestedTransportPropertyNames.length; i++) {
+                if (keyValue[0].equalsIgnoreCase(requestedTransportPropertyNames[i])) {
+                    headersArray[i] = keyValue[1];
+                }
+            }
+        }
+        for (int i = 0; i < requestedTransportPropertyNames.length; i++) {
+            if (headersArray[i] == null || headersArray[i].equalsIgnoreCase(GrpcConstants.EMPTY_STRING)) {
+                throw new SiddhiAppRuntimeException("Missing header " + requestedTransportPropertyNames[i]);
+            }
+        }
+        return headersArray;
     }
 }
